move size.go into d2geom
