+++ d2common/d2geom/size.go
@@ -1,4 +1,4 @@
-package d2common
+package d2geom
 
 // Size represents a size
 type Size struct {
